修复功法快照真元成本精度
将功法 qi_cost 从 registry 到 protobuf 与客户端模型全链路保留为 f64，并用大数边界回归锁定服务端扣费与客户端展示一致。

Model: claude-haiku-4-5-20251001

+++ client/src/main/java/com/bong/client/network/TechniquesSnapshotHandler.java
@@ -75,7 +75,7 @@ public final class TechniquesSnapshotHandler implements ServerDataHandler {
         boolean active = readBoolean(obj, "active", false);
         String description = SkillBarConfigHandler.readString(obj, "description");
         String requiredRealm = SkillBarConfigHandler.readString(obj, "required_realm");
-        float qiCost = (float) readDouble(obj, "qi_cost", 0.0);
+        double qiCost = readDouble(obj, "qi_cost", 0.0);
         int castTicks = (int) SkillBarConfigHandler.readLong(obj, "cast_ticks", 0L);
         int cooldownTicks = (int) SkillBarConfigHandler.readLong(obj, "cooldown_ticks", 0L);
         float range = (float) readDouble(obj, "range", 0.0);
